Free Energy Foundation
Why this?

Free Software today is like a king sized bed. What started on the 386, became a kernel with over 8 million lines of code. Today you can use Linux on microcontrollers, 64 core processors, but nothing user space with a priority on energy-first design. The purpose of this repository is to explicitly focus on designing a kernel based for a solar powerable mobile device. https://hackaday.io/project/177716-the-libre-autarkic-laptop
This project rejects the mainline kernel philosophy and seeks to make an RTOS such as Zephyr for userspace applications.